+++ .github/workflows/scripts/win-test-services.ps1
@@ -10,10 +10,25 @@ param (
 $ErrorActionPreference = 'Stop'
 Set-PSDebug -Trace 1
 
-function check_regkey([string]$name, [string]$value) {
-    $actual = Get-ItemPropertyValue -PATH "HKLM:\SYSTEM\CurrentControlSet\Control\Session Manager\Environment" -name "$name"
-    if ( "$value" -ne "$actual" ) {
-        throw "Environment variable $name is not properly set. Found: '$actual', Expected '$value'"
+function check_collector_svc_environment([hashtable]$expected_env_vars) {
+    $actual_env_vars = @{}
+    try {
+        $env_array = Get-ItemPropertyValue -Path "HKLM:\SYSTEM\CurrentControlSet\Services\splunk-otel-collector" -Name "Environment"
+        foreach ($entry in $env_array) {
+            $key, $value = $entry.Split("=", 2)
+            $actual_env_vars.Add($key, $value)
+        }
+    } catch {
+        Write-Host "Assuming an old version of the collector with environment variables at the machine scope"
+        $actual_env_vars = [Environment]::GetEnvironmentVariables("Machine")<#Do this if a terminating exception happens#>
+    }
+
+    foreach ($key in $expected_env_vars.Keys) {
+        $expected_value = $expected_env_vars[$key]
+        $actual_value = $actual_env_vars[$key]
+        if ($expected_value -ne $actual_value) {
+            throw "Environment variable $key is not properly set. Found: '$actual_value', Expected '$expected_value'"
+        }
     }
 }
 
@@ -24,16 +39,18 @@ function service_running([string]$name) {
 $api_url = "https://api.${realm}.signalfx.com"
 $ingest_url = "https://ingest.${realm}.signalfx.com"
 
-check_regkey -name "SPLUNK_CONFIG" -value "${env:PROGRAMDATA}\Splunk\OpenTelemetry Collector\${mode}_config.yaml"
-check_regkey -name "SPLUNK_ACCESS_TOKEN" -value "$access_token"
-check_regkey -name "SPLUNK_REALM" -value "$realm"
-check_regkey -name "SPLUNK_API_URL" -value "$api_url"
-check_regkey -name "SPLUNK_INGEST_URL" -value "$ingest_url"
-check_regkey -name "SPLUNK_TRACE_URL" -value "${ingest_url}/v2/trace"
-check_regkey -name "SPLUNK_HEC_URL" -value "${ingest_url}/v1/log"
-check_regkey -name "SPLUNK_HEC_TOKEN" -value "$access_token"
-check_regkey -name "SPLUNK_BUNDLE_DIR" -value "${env:PROGRAMFILES}\Splunk\OpenTelemetry Collector\agent-bundle"
-check_regkey -name "SPLUNK_MEMORY_TOTAL_MIB" -value  "$memory"
+check_collector_svc_environment @{
+  "SPLUNK_CONFIG"           = "${env:PROGRAMDATA}\Splunk\OpenTelemetry Collector\${mode}_config.yaml";
+  "SPLUNK_ACCESS_TOKEN"     = "$access_token";
+  "SPLUNK_REALM"            = "$realm";
+  "SPLUNK_API_URL"          = "$api_url";
+  "SPLUNK_INGEST_URL"       = "$ingest_url";
+  "SPLUNK_TRACE_URL"        = "${ingest_url}/v2/trace";
+  "SPLUNK_HEC_URL"          = "${ingest_url}/v1/log";
+  "SPLUNK_HEC_TOKEN"        = "$access_token";
+  "SPLUNK_BUNDLE_DIR"       = "${env:PROGRAMFILES}\Splunk\OpenTelemetry Collector\agent-bundle";
+  "SPLUNK_MEMORY_TOTAL_MIB" = "$memory";
+}
 
 if ((service_running -name "splunk-otel-collector")) {
     write-host "splunk-otel-collector service is running."